baixando a imagem na viewmodel
